+++ src/pland/hooks/listeners/EntityListeners.ll.cc
@@ -1,11 +1,14 @@
 
+#include "mc/world/actor/ActorType.h"
 #include "pland/hooks/EventListener.h"
 #include "pland/hooks/listeners/ListenerHelper.h"
 
 #include "ll/api/event/EventBus.h"
+#include "ll/api/event/entity/ActorHurtEvent.h"
 #include "ll/api/event/world/SpawnMobEvent.h"
 
 #include "mc/server/ServerPlayer.h"
+#include "mc/world/level/Level.h"
 
 #include "pland/PLand.h"
 #include "pland/infra/Config.h"
@@ -33,6 +36,41 @@ void EventListener::registerLLEntityListeners() {
                 if (!tab.allowMonsterSpawn) mob->despawn();
             } else {
                 if (!tab.allowAnimalSpawn) mob->despawn();
+            }
+        });
+    });
+
+    RegisterListenerIf(Config::cfg.listeners.ActorHurtEvent, [&]() {
+        return bus->emplaceListener<ll::event::ActorHurtEvent>([db](ll::event::ActorHurtEvent& ev) {
+            auto& actor  = ev.self();
+            auto& source = ev.source();
+
+            bool const isPlayerDamage = source.getEntityType() == ActorType::Player;
+            if (!isPlayerDamage) {
+                return; // skip non-player damage
+            }
+
+            auto sourcePlayer = actor.getLevel().getPlayer(source.getEntityUniqueID());
+            if (!sourcePlayer) {
+                return;
+            }
+
+            auto land = db->getLandAt(actor.getPosition(), actor.getDimensionId());
+            if (PreCheckLandExistsAndPermission(land, sourcePlayer->getUuid())) return;
+
+            auto const& typeName = actor.getTypeName();
+            auto const& tab      = land->getPermTable();
+
+            if (isPlayerDamage) {
+                CANCEL_AND_RETURN_IF(!tab.allowPlayerDamage);
+            } else if (Config::cfg.protection.mob.hostileMobTypeNames.contains(typeName)) {
+                CANCEL_AND_RETURN_IF(!tab.allowMonsterDamage);
+            } else if (Config::cfg.protection.mob.specialMobTypeNames.contains(typeName)) {
+                CANCEL_AND_RETURN_IF(!tab.allowSpecialDamage);
+            } else if (Config::cfg.protection.mob.passiveMobTypeNames.contains(typeName)) {
+                CANCEL_AND_RETURN_IF(!tab.allowPassiveDamage);
+            } else if (Config::cfg.protection.mob.customSpecialMobTypeNames.contains(typeName)) {
+                CANCEL_AND_RETURN_IF(!tab.allowCustomSpecialDamage);
             }
         });
     });