+++ src/pland/infra/Config.h
@@ -18,7 +18,7 @@ struct ForbiddenRange {
 };
 
 struct Config {
-    int              version{25};
+    int              version{26};
     ll::io::LogLevel logLevel{ll::io::LogLevel::Info};
 
     EconomyConfig economy;
@@ -114,6 +114,7 @@ struct Config {
         bool EndermanLeaveBlockEvent{true};                   // 末影人放下方块
         bool EndermanTakeBlockEvent{true};                    // 末影人拿走方块
         bool DragonEggBlockTeleportBeforeEvent{true};         // 龙蛋传送
+        bool PlayerUseItemEvent{true};                        // 玩家使用物品
     } listeners;
 
     struct {