+++ CHANGELOG.md
@@ -7,9 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### 🐛 问题修复
+### 🐛 已知问题修复
 
-- 修复领地外无权限玩家可击杀领地内实体 #140 @engsr6982
+- [#140] 领地外无权限玩家可击杀领地内实体  
+  依赖 LeviLamina ≥ 1.5.2，低版本仍可能复现。 @engsr6982
 
 ### 🧩 逻辑优化