+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### 🧹 其他改动
 
 - 移除 `debug_shape` 相关封装，改为外部独立 **DebugShape.dll** 组件
+- 适配 Levilamina v1.5.1 @yangyangzhong82
 
 ## [0.13.0] - 2025-8-27