+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### 🐛 问题修复
 
 - [#149] [#147] 修复 Mod 关闭时事件监听器未卸载导致的报错 @engsr6982
+- [#150] 修复伤害拦截异常 @engsr6982
 
 ### 🧹 其他改动